CG-92: initial attempt at scheduler and provisional unit tests.

+++ core/logging/messages.hpp
@@ -86,7 +86,7 @@ struct __attribute__((packed)) GPSLogEntry {
 			int32_t    velE;
 			int32_t    velD;
 			int32_t    gSpeed;
-			int32_t    headMot;
+			double     headMot;
 			uint32_t   sAcc;
 			uint32_t   headAcc;
 			uint16_t   pDOP;
